Add MkDocs configuration and update website layout
A MkDocs configuration was added allowing MkDocs to generate project documentation. This config file is placed under source control for changes to be tracked. Also, the 'Open Help' link in the website layout now points to this new MkDocs generated documentation. Lastly, the MkDocs package was added to the project dependencies.