Update configuration handling and references
This commit updates the way application and secret configurations are handled in "kringlecraft/app.py". Changes include the addition of secret configuration to the json file, and updates to the way both application and secret configurations are referenced. Furthermore, minor changes are made to templating in "_layout.html" to reflect structural changes in configuration.