Refactor border styling in shared layout
Replaced inline border styling in 'navbar' and 'container-fluid' elements in the shared layout with a new class 'c-border' defined in main.css. The HTML is now cleaner and the reused border style is managed conveniently in the CSS file for consistent UI across the website. The new update allows easier style changes in the future.