Add statistics view and user count functionality
A new statistics view has been added to the application, accessible from the main navigation. This view displays statistics about the elements stored in the database. To support this, a new method `get_user_count` has been added to `user_services.py` to count the number of users. The user interface has been updated to include a link to the new statistics page.