Add functionality to view individual user profile
In this commit, a new feature was added that allows individual user profiles to be viewed. This was achieved by creating a new page called 'user.html' and creating relevant routes and methods in the backend. Additionally, a change was made to now lookup only active users during login, utilizing a new method 'find_active_user_by_id'.