Add detailed view for individual World
A new route and associated functionality were added for the display of a specific 'World'. This includes creation of a new form and services for a specific World, along with error handling. Also, adjusted file saving routines to use a new helper function to build file paths. Introduced dedicated error pages for non-existent user and world.