Add objective view and related functionality
This commit adds a new view (objectives) to the application along with all related functions like creating, editing, or deleting an objective. These functions are available in the ObjectiveForm. The information about objectives is displayed on several pages, the related details are added in their respective HTML templates. Also, objectives database model and services are introduced. Overall, this commit provides a complete implementation for handling objectives within the application.