Process form data and refactor delete operations
Added process method call for WorldForm, RoomForm, and ObjectiveForm to process form data. Refactoring was done on delete operations in the data_views file- the initial form data setup was replaced with the actual operations to delete a world, room, and objective. The code now better reflects the actions being taken.