Add image handling and display functionalities
Functions for handling images such as uploading and retrieving are added to the system. In addition, the feature to display these images in the challenge editor of the application is implemented ensuring a smoother user experience. Also, error handling methods are modified to return JSON responses. CSS for displaying images is also updated.