Add solution display to objective page
The data view for objectives now imports and uses a function to find active solutions. This solution content has been added to the objective's template page, allowing user to view the solution for each objective. A function to retrieve active solutions from the database has also been added to the services module.