Add user services and display all solutions in objectives
The user services module has been imported, and new function to get a dictionary of all users is added. Changes were also made in the objective page to now display all solutions, if available, instead of just one for an objective. A new variable 'candidate_solutions' is added to contain all solutions for an objective. The template for objectives is updated to include a section listing all solutions along with the user who proposed them.