Add progress page and linking functionality
A new progress page is introduced in the task directory and this page is filled with the relevant details of progress for a specific 'world'. This page can now be navigated to via the worlds' buttons. Changes to the world.html template are made such that the 'progress' button now links to this progress page, carrying necessary id parameters.