Add highlight feature for data views
Included a 'highlight' parameter in the form processing of data_views.py to visually distinguish specific data in the display. This was implemented for worlds, rooms, and objectives. This parameter is passed to templates through rendering, where corresponding HTML has been updated to reflect the optional highlight styling. A new CSS class was added for this styling purpose.