Replace numeric role codes with constants in application
This commit includes the creation of a new constants file that holds user role codes, which are then used to replace numeric role codes throughout the application. This change improves code readability and maintainability by giving explicit names to the role codes. Additionally, the minimum length constraint for the search string field was implemented and context size in the search_binary_text method was expanded.